backend/beam_requirements.txt
# Beam-specific requirements for ML pipeline processing
# These are the heavy dependencies that will run on Beam GPU instances

# Core ML and AI dependencies
voyageai==0.3.3
openai==1.97.0
langchain==0.3.26
langchain-openai==0.3.28
langsmith==0.4.8

# PDF processing (heavy dependencies)
PyMuPDF==1.26.3
pypdf==5.8.0
pdf2image==1.17.0

# Advanced PDF processing (for hybrid partition system)
# Full PDF functionality needed for OCR and hi-res processing
unstructured[pdf]==0.10.25
nltk==3.8.1

# Scientific computing
numpy==1.26.4
scipy==1.16.0
pandas==2.3.1
matplotlib==3.8.2

# Database and HTTP
supabase==2.17.0
psycopg2-binary==2.9.9
httpx==0.28.1
websockets==15.0.1

# Configuration and validation
pydantic[email]==2.11.7
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==24.1.0
psutil==5.9.8

# Logging
structlog==23.2.0
rich==13.7.0

# Beam SDK - Not needed in requirements, Beam handles this automatically
# beam-sdk 